+++ demo/api/src/app.module.ts
@@ -1,5 +1,4 @@
 import {
-    AuthModule,
     BlobStorageConfig,
     BlobStorageModule,
     BlocksModule,
@@ -8,10 +7,11 @@ import {
     BuildsModule,
     ContentScope,
     ContentScopeModule,
+    CurrentUserInterface,
     DamModule,
     FilesService,
-    GlobalAuthGuard,
     ImagesService,
+    KubernetesModule,
     PageTreeModule,
     PageTreeService,
     PublicUploadModule,
@@ -20,7 +20,6 @@ import {
 import { ApolloDriver } from "@nestjs/apollo";
 import { Module } from "@nestjs/common";
 import { ConfigType } from "@nestjs/config";
-import { APP_GUARD } from "@nestjs/core";
 import { GraphQLModule } from "@nestjs/graphql";
 import { ConfigModule } from "@src/config/config.module";
 import { configNS } from "@src/config/config.namespace";
@@ -30,7 +29,7 @@ import { PagesModule } from "@src/pages/pages.module";
 import { PredefinedPage } from "@src/predefined-page/entities/predefined-page.entity";
 import { Request } from "express";
 
-import { CurrentUserLoaderService } from "./auth/current-user-loader.service";
+import { AuthModule } from "./auth/auth.module";
 import { FooterModule } from "./footer/footer.module";
 import { Link } from "./links/entities/link.entity";
 import { MenusModule } from "./menus/menus.module";
@@ -66,23 +65,9 @@ import { RedirectScope } from "./redirects/dto/redirect-scope";
             }),
             inject: [configNS.KEY, BlocksTransformerService],
         }),
-        AuthModule.registerAsync({
-            imports: [ConfigModule],
-            useFactory: async (config: ConfigType<typeof configNS>) => ({
-                config: {
-                    apiPassword: config.API_PASSWORD,
-                    idpConfig: {
-                        url: config.IDP_API_URL,
-                        password: config.IDP_API_PASSWORD,
-                        clientId: config.IDP_CLIENT_ID,
-                    },
-                },
-            }),
-            inject: [configNS.KEY],
-            currentUserLoaderService: CurrentUserLoaderService,
-        }),
+        AuthModule,
         ContentScopeModule.forRoot({
-            canAccessScope(requestScope: ContentScope, user) {
+            canAccessScope(requestScope: ContentScope, user: CurrentUserInterface) {
                 if (!user.domains) return true; //all domains
                 return user.domains.includes(requestScope.domain);
             },
@@ -100,7 +85,7 @@ import { RedirectScope } from "./redirects/dto/redirect-scope";
             },
             inject: [PageTreeService, FilesService, ImagesService],
         }),
-        BuildsModule.registerAsync({
+        KubernetesModule.registerAsync({
             imports: [ConfigModule],
             useFactory: async (config: ConfigType<typeof configNS>) => ({
                 config: {
@@ -109,6 +94,7 @@ import { RedirectScope } from "./redirects/dto/redirect-scope";
             }),
             inject: [configNS.KEY],
         }),
+        BuildsModule,
         LinksModule,
         PagesModule,
         PageTreeModule.forRoot({
@@ -195,6 +181,5 @@ import { RedirectScope } from "./redirects/dto/redirect-scope";
         PredefinedPageModule,
         ProductsModule,
     ],
-    providers: [{ provide: APP_GUARD, useClass: GlobalAuthGuard }],
 })
 export class AppModule {}