+++ packages/admin/blocks-admin/src/blocks/factories/createBlocksBlock.tsx
@@ -188,14 +188,24 @@ export function createBlocksBlock({
             };
         },
         isValid: async (state) =>
-            await parallelAsyncEvery(state.blocks, async (c) => {
+            parallelAsyncEvery(state.blocks, async (c) => {
                 const block = blockForType(c.type);
                 if (!block) {
                     throw new Error(`No Block found for type ${c.type}`); // for TS
                 }
-                return await block.isValid(c.props);
+                return block.isValid(c.props);
             }),
         childBlockCount: (state) => state.blocks.length,
+
+        anchors: (state) => {
+            return state.blocks.reduce<string[]>((anchors, child) => {
+                const block = blockForType(child.type);
+                if (!block) {
+                    throw new Error(`No Block found for type ${child.type}`); // for TS
+                }
+                return [...anchors, ...(block.anchors?.(child.props) ?? [])];
+            }, []);
+        },
 
         definesOwnPadding: true,