+++ packages/admin/blocks-admin/src/iframebridge/IFrameMessage.ts
@@ -65,7 +65,6 @@ export enum AdminMessageType {
     Block = "Block",
     SelectComponent = "SelectComponent",
     HoverComponent = "HoverComponent",
-    AccessToken = "AccessToken",
 }
 export interface IAdminBlockMessage {
     cometType: AdminMessageType.Block;
@@ -89,9 +88,4 @@ export interface IAdminHoverComponentMessage {
     };
 }
 
-export interface IAdminAccessTokenMessage {
-    cometType: AdminMessageType.AccessToken;
-    accessToken: string;
-}
-
-export type AdminMessage = IAdminBlockMessage | IAdminSelectComponentMessage | IAdminHoverComponentMessage | IAdminAccessTokenMessage;
+export type AdminMessage = IAdminBlockMessage | IAdminSelectComponentMessage | IAdminHoverComponentMessage;