+++ packages/admin/cms-admin/src/index.ts
@@ -1,3 +1,4 @@
+export { AnchorBlock } from "./blocks/AnchorBlock";
 export { CmsBlockContext, CmsBlockContextProvider } from "./blocks/CmsBlockContextProvider";
 export { createImageLinkBlock } from "./blocks/createImageLinkBlock";
 export { createLinkBlock } from "./blocks/createLinkBlock";
@@ -19,7 +20,6 @@ export { BuildRuntime } from "./builds/BuildRuntime";
 export { Publisher } from "./builds/Publisher";
 export { includeInvisibleContentContext } from "./common/apollo/links/includeInvisibleContentContext";
 export { DropdownMenuItem } from "./common/DropdownMenuItem";
-export { AuthorizationErrorPage } from "./common/errors/AuthorizationErrorPage";
 export { BuildInformationProvider } from "./common/header/about/build-information/BuildInformationProvider";
 export { useBuildInformation } from "./common/header/about/build-information/useBuildInformation";
 export { Header } from "./common/header/Header";