+++ packages/api/cms-api/src/builds/build-templates.service.ts
@@ -1,22 +1,22 @@
+import { V1CronJob } from "@kubernetes/client-node";
 import { Injectable } from "@nestjs/common";
 
+import { CurrentUserInterface } from "../auth/current-user/current-user";
+import { ContentScopeService } from "../content-scope/content-scope.service";
+import { KubernetesService } from "../kubernetes/kubernetes.service";
 import { BUILDER_LABEL, INSTANCE_LABEL } from "./builds.constants";
-import { BuildTemplateObject } from "./dto/build-template.object";
-import { KubernetesService } from "./kubernetes.service";
 
 @Injectable()
 export class BuildTemplatesService {
-    constructor(private readonly kubernetesService: KubernetesService) {}
+    constructor(private readonly kubernetesService: KubernetesService, private readonly contentScopeService: ContentScopeService) {}
 
-    async getBuildTemplates(): Promise<BuildTemplateObject[]> {
-        if (this.kubernetesService.localMode) {
-            throw Error("Not available in local mode!");
-        }
-
-        const builderCronJobs = await this.kubernetesService.getAllCronJobs(
+    async getAllowedBuilderCronJobs(user: CurrentUserInterface): Promise<V1CronJob[]> {
+        const allCronJobs = await this.kubernetesService.getAllCronJobs(
             `${BUILDER_LABEL} = true, ${INSTANCE_LABEL} = ${this.kubernetesService.helmRelease}`,
         );
 
-        return builderCronJobs.map((cronJob) => ({ id: cronJob.metadata?.uid as string, name: cronJob.metadata?.name as string }));
+        return allCronJobs.filter((cronJob) => {
+            return this.contentScopeService.canAccessScope(this.kubernetesService.getContentScope(cronJob), user);
+        });
     }
 }