+++ packages/api/cms-api/src/dam/files/files.resolver.ts
@@ -131,7 +131,7 @@ export class FilesResolver {
 
     @ResolveField(() => String)
     async damPath(@Parent() file: File): Promise<string> {
-        return await this.filesService.getDamPath(file);
+        return this.filesService.getDamPath(file);
     }
 
     @ResolveField(() => [BlockIndexDependency])