+++ CHANGELOG.md
@@ -2,7 +2,54 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## NEXT
+## Next major
+
+### Highlights
+
+-   Add `AnchorBlock` to support linking to anchors
+
+### @comet/blocks-admin
+
+#### Changes
+
+-   Add `anchors` method to `BlockInterface` for a block to specify its anchors
+
+### @comet/cms-admin
+
+-   Breaking: Remove dependency @comet/react-app-auth and all occurences
+
+#### Changes
+
+-   Add `AnchorBlock`
+-   Add support for anchors to `InternalLinkBlock`
+-   Add `anchors` method to `DocumentInterface` for a document to specify its anchors
+
+### @comet/cms-api
+
+-   Restrict access to builds based on `ContentScopeModule`
+-   Decouple KubernetesModule from BuildsModule (Breaking: BuildsModuleConfig is now KubernetesModuleConfig)
+-   Breaking: AuthModule removes support for Bearer-Token
+
+#### Changes
+
+-   Add `AnchorBlock`
+-   Add support for anchors to `InternalLinkBlock`
+
+### @comet/cms-site
+
+-   Breaking: Remove dependency next-auth and all occurences
+-   access-token-service-worker.js in not supported anymore, can be removed
+
+#### Changes
+
+-   Add `AnchorBlock`
+-   Add support for anchors to `InternalLinkBlock`
+
+### @comet/eslint-config
+
+-   Enable [no-return-await](https://eslint.org/docs/latest/rules/no-return-await)
+
+## Next minor
 
 ### @comet/admin